requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
awesome-slugify==1.6.5
awscli==1.11.36
boto==2.45.0
botocore==1.4.93          # via awscli, s3transfer
cffi==1.9.1               # via cryptography
colorama==0.3.7           # via awscli
cryptography==1.7.1       # via paramiko
docutils==0.13.1          # via awscli, botocore
enum34==1.1.6             # via cryptography
fabric==1.13.1
futures==3.0.5            # via s3transfer
idna==2.2                 # via cryptography
ipaddress==1.0.17         # via cryptography
jmespath==0.9.0           # via botocore
ordereddict==1.1
paramiko==2.1.1           # via fabric
pyasn1==0.1.9             # via cryptography, paramiko, rsa
pycparser==2.17           # via cffi
python-dateutil==2.6.0    # via botocore
python-frontmatter==0.3.1
pyyaml==3.12
regex==2016.12.27         # via awesome-slugify
rsa==3.4.2                # via awscli
s3transfer==0.1.10        # via awscli
six==1.10.0               # via cryptography, python-dateutil, python-frontmatter
unidecode==0.4.19         # via awesome-slugify

# The following packages are considered to be unsafe in a requirements file:
# setuptools                # via cryptography
